refactor: alteração na organização
