+++ src/main/java/com/AgiBank/dao/contribuicao/ContribuicaoDAO.java
@@ -1,0 +1,5 @@
+package com.AgiBank.dao.contribuicao;
+
+public interface ContribuicaoDAO {
+    //verificar usuário
+}
